flask_app/requirements.txt
apifairy==0.9.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.28.0
flask_migrate==3.1.0
psycopg2 == 2.9.3
rich == 12.5.1
gunicorn==20.1.0
uvicorn==0.18.2
Flask-JWT-Extended==4.4.2
redis==4.3.4
gevent==21.12.0
pydantic==1.9.1
click==8.1.3
